AdventOfCode.Tests/Year2021/Day06/ProblemTests.cs
using AdventOfCode.Year2021.Day06;
using NUnit.Framework;
using System.Text;

namespace AdventOfCode.Tests.Year2021.Day06;

[TestFixture]
public class ProblemTests
{
  [OneTimeSetUp]
  public void SetUp()
  {
    StringBuilder inputBuilder = new();
    inputBuilder.AppendLine("3,4,3,1,2");

    _input = inputBuilder.ToString();
  }

  private string _input = string.Empty;

  [Test]
  public void Part1Example()
  {
    // arrange
    Problem problem = new();

    // act
    string result = problem.Part1(_input);

    // assert
    const string expected = "5934";
    Assert.That(result, Is.EqualTo(expected));
  }

  [Test]
  public void Part2Example()
  {
    // arrange
    Problem problem = new();
    
    // act
    string result = problem.Part2(_input);
    
    // assert
    const string expected = "26984457539";
    Assert.That(result, Is.EqualTo(expected));
  }
}
